requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.4

# Database
aiosqlite==0.20.0

# OCR - Free options
pytesseract==0.3.13
easyocr==1.7.2
pillow==11.0.0

# OCR - AI Vision (optional, for better accuracy)
anthropic==0.40.0
openai==1.58.1
requests==2.32.3

# OCR - Local vision LLM (build from source with CUDA, see build_llama_cuda.bat)
llama-cpp-python==0.3.16

# PDF handling
pypdf2==3.0.1
pdf2image==1.17.0

# Dropbox
dropbox==12.0.2

# Email
aiosmtplib==3.0.2

# Configuration
python-dotenv==1.0.1

# Development
ruff==0.8.4
pytest==8.3.4
pytest-asyncio==0.25.0
